+++ resources/assets/photonCms/core/js/components/Generator/ModuleConfigurator/ModuleConfigurator.moduleOptions.js
@@ -171,7 +171,7 @@ export const getModuleOptions = (ModuleConfigurator) => {
             mutation: 'generator/UPDATE_GENERATOR_NON_GROUPED_FIELDS_TO_BOTTOM',
             name: 'module[non_grouped_to_bottom]',
             tooltip: 'If checked, non-grouped fields will be rendered below the grouped fields instead of above.',
-            value: _.has(selectedModule, 'non_grouped_to_bottom') ? selectedModule.non_grouped_to_bottom : 0,
+            value: _.has(selectedModule, 'non_grouped_to_bottom') ? selectedModule.non_grouped_to_bottom : false,
             vueComponent: 'Boolean',
         },
     ];